maint: move RegridToZooms, FilledImageLike and DemeanImage into nibabel module
