+++ niworkflows/interfaces/images.py
@@ -719,58 +719,6 @@ class ValidateImage(SimpleInterface):
         return runtime
 
 
-class _DemeanImageInputSpec(BaseInterfaceInputSpec):
-    in_file = File(exists=True, mandatory=True, desc="image to be demeaned")
-    in_mask = File(
-        exists=True, mandatory=True, desc="mask where median will be calculated"
-    )
-    only_mask = traits.Bool(False, usedefault=True, desc="demean only within mask")
-
-
-class _DemeanImageOutputSpec(TraitedSpec):
-    out_file = File(exists=True, desc="demeaned image")
-
-
-class DemeanImage(SimpleInterface):
-    input_spec = _DemeanImageInputSpec
-    output_spec = _DemeanImageOutputSpec
-
-    def _run_interface(self, runtime):
-        self._results["out_file"] = demean(
-            self.inputs.in_file,
-            self.inputs.in_mask,
-            only_mask=self.inputs.only_mask,
-            newpath=runtime.cwd,
-        )
-        return runtime
-
-
-class _FilledImageLikeInputSpec(BaseInterfaceInputSpec):
-    in_file = File(exists=True, mandatory=True, desc="image to be demeaned")
-    fill_value = traits.Float(1.0, usedefault=True, desc="value to fill")
-    dtype = traits.Enum(
-        "float32", "uint8", usedefault=True, desc="force output data type"
-    )
-
-
-class _FilledImageLikeOutputSpec(TraitedSpec):
-    out_file = File(exists=True, desc="demeaned image")
-
-
-class FilledImageLike(SimpleInterface):
-    input_spec = _FilledImageLikeInputSpec
-    output_spec = _FilledImageLikeOutputSpec
-
-    def _run_interface(self, runtime):
-        self._results["out_file"] = nii_ones_like(
-            self.inputs.in_file,
-            self.inputs.fill_value,
-            self.inputs.dtype,
-            newpath=runtime.cwd,
-        )
-        return runtime
-
-
 class _MatchHeaderInputSpec(BaseInterfaceInputSpec):
     reference = File(
         exists=True, mandatory=True, desc="NIfTI file with reference header"
@@ -865,42 +813,6 @@ def normalize_xform(img):
     new_img.set_sform(xform, xform_code)
     new_img.set_qform(xform, xform_code)
     return new_img
-
-
-def demean(in_file, in_mask, only_mask=False, newpath=None):
-    """Demean ``in_file`` within the mask defined by ``in_mask``."""
-    import os
-    import numpy as np
-    import nibabel as nb
-    from nipype.utils.filemanip import fname_presuffix
-
-    out_file = fname_presuffix(in_file, suffix="_demeaned", newpath=os.getcwd())
-    nii = nb.load(in_file)
-    msk = np.asanyarray(nb.load(in_mask).dataobj)
-    data = nii.get_fdata()
-    if only_mask:
-        data[msk > 0] -= np.median(data[msk > 0])
-    else:
-        data -= np.median(data[msk > 0])
-    nb.Nifti1Image(data, nii.affine, nii.header).to_filename(out_file)
-    return out_file
-
-
-def nii_ones_like(in_file, value, dtype, newpath=None):
-    """Create a NIfTI file filled with ``value``, matching properties of ``in_file``."""
-    import os
-    import numpy as np
-    import nibabel as nb
-
-    nii = nb.load(in_file)
-    data = np.ones(nii.shape, dtype=float) * value
-
-    out_file = os.path.join(newpath or os.getcwd(), "filled.nii.gz")
-    nii = nb.Nifti1Image(data, nii.affine, nii.header)
-    nii.set_data_dtype(dtype)
-    nii.to_filename(out_file)
-
-    return out_file
 
 
 class _SignalExtractionInputSpec(BaseInterfaceInputSpec):
