+++ niworkflows/interfaces/nibabel.py
@@ -176,3 +176,111 @@ class MergeSeries(SimpleInterface):
 
         self._results["out_file"] = out_file
         return runtime
+
+
+class _RegridToZoomsInputSpec(BaseInterfaceInputSpec):
+    in_file = File(
+        exists=True, mandatory=True, desc="a file whose resolution is to change"
+    )
+    zooms = traits.Tuple(
+        traits.Float,
+        traits.Float,
+        traits.Float,
+        mandatory=True,
+        desc="the new resolution",
+    )
+    order = traits.Int(3, usedefault=True, desc="order of interpolator")
+    clip = traits.Bool(
+        True,
+        usedefault=True,
+        desc="clip the data array within the original image's range",
+    )
+    smooth = traits.Either(
+        traits.Bool(),
+        traits.Float(),
+        default=False,
+        usedefault=True,
+        desc="apply gaussian smoothing before resampling",
+    )
+
+
+class _RegridToZoomsOutputSpec(TraitedSpec):
+    out_file = File(exists=True, dec="the regridded file")
+
+
+class RegridToZooms(SimpleInterface):
+    """Change the resolution of an image (regrid)."""
+
+    input_spec = _RegridToZoomsInputSpec
+    output_spec = _RegridToZoomsOutputSpec
+
+    def _run_interface(self, runtime):
+        from ..utils.images import resample_by_spacing
+
+        self._results["out_file"] = fname_presuffix(
+            self.inputs.in_file, suffix="_regrid", newpath=runtime.cwd
+        )
+        resample_by_spacing(
+            self.inputs.in_file,
+            self.inputs.zooms,
+            order=self.inputs.order,
+            clip=self.inputs.clip,
+            smooth=self.inputs.smooth,
+        ).to_filename(self._results["out_file"])
+        return runtime
+
+
+class _DemeanImageInputSpec(BaseInterfaceInputSpec):
+    in_file = File(exists=True, mandatory=True, desc="image to be demeaned")
+    in_mask = File(
+        exists=True, mandatory=True, desc="mask where median will be calculated"
+    )
+    only_mask = traits.Bool(False, usedefault=True, desc="demean only within mask")
+
+
+class _DemeanImageOutputSpec(TraitedSpec):
+    out_file = File(exists=True, desc="demeaned image")
+
+
+class DemeanImage(SimpleInterface):
+    input_spec = _DemeanImageInputSpec
+    output_spec = _DemeanImageOutputSpec
+
+    def _run_interface(self, runtime):
+        from ..utils.images import demean
+
+        self._results["out_file"] = demean(
+            self.inputs.in_file,
+            self.inputs.in_mask,
+            only_mask=self.inputs.only_mask,
+            newpath=runtime.cwd,
+        )
+        return runtime
+
+
+class _FilledImageLikeInputSpec(BaseInterfaceInputSpec):
+    in_file = File(exists=True, mandatory=True, desc="image to be demeaned")
+    fill_value = traits.Float(1.0, usedefault=True, desc="value to fill")
+    dtype = traits.Enum(
+        "float32", "uint8", usedefault=True, desc="force output data type"
+    )
+
+
+class _FilledImageLikeOutputSpec(TraitedSpec):
+    out_file = File(exists=True, desc="demeaned image")
+
+
+class FilledImageLike(SimpleInterface):
+    input_spec = _FilledImageLikeInputSpec
+    output_spec = _FilledImageLikeOutputSpec
+
+    def _run_interface(self, runtime):
+        from ..utils.images import nii_ones_like
+
+        self._results["out_file"] = nii_ones_like(
+            self.inputs.in_file,
+            self.inputs.fill_value,
+            self.inputs.dtype,
+            newpath=runtime.cwd,
+        )
+        return runtime
